Add files via upload
I updated it to a new version. In this notebook, I cleaned it up a little bit and updated the font to a-avenir-next-lt-pro using a new package. For the code, I simplified three plots to one so the font format could apply to all text, and the output would be cleaner. For the plot, I updated the bar color and minor grid color to the required color from the style guide and removed the border line, changed the rotation degree to 65 to meet the requirement. I also updated a brief style guide to the notebook and full version of style guide on repo, we can use both as guideline to modify the graph.