Deleting Sarah's old code
Continuing to clean up the repo. I don't think we need to keep this old script (made by Sarah) that serves as proof of concept for importing Style.R using source().

Do you agree that it's best to get rid of this in order to have a cleaner repo? Or do you think it's a helpful walkthrough for any future policy analysts who don't have strong coding abilities and are trying to learn how the repo/scripts work?